JWT + fine-grained API access control

+++ step-7/src/main/java/io/vertx/guides/wiki/http/HttpServerVerticle.java
@@ -29,6 +29,9 @@ import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.auth.AuthProvider;
+import io.vertx.ext.auth.User;
+import io.vertx.ext.auth.jwt.JWTAuth;
+import io.vertx.ext.auth.jwt.JWTOptions;
 import io.vertx.ext.auth.shiro.ShiroAuth;
 import io.vertx.ext.auth.shiro.ShiroAuthOptions;
 import io.vertx.ext.auth.shiro.ShiroAuthRealmType;
@@ -110,7 +113,48 @@ public class HttpServerVerticle extends AbstractVerticle {
         .end();
     });
 
+    JWTAuth jwtAuth = JWTAuth.create(vertx, new JsonObject()
+      .put("keyStore", new JsonObject()
+        .put("path", "keystore.jceks")
+        .put("type", "jceks")
+        .put("password", "secret")));
+
     Router apiRouter = Router.router(vertx);
+
+    apiRouter.route().handler(JWTAuthHandler.create(jwtAuth, "/api/token"));
+
+    apiRouter.get("/token").handler(context -> {
+
+      JsonObject creds = new JsonObject()
+        .put("username", context.request().getHeader("login"))
+        .put("password", context.request().getHeader("password"));
+      auth.authenticate(creds, authResult -> {
+
+        if (authResult.succeeded()) {
+          User user = authResult.result();
+          user.isAuthorised("create", canCreate -> {
+            user.isAuthorised("delete", canDelete -> {
+              user.isAuthorised("update", canUpdate -> {
+
+                String token = jwtAuth.generateToken(
+                  new JsonObject()
+                    .put("username", context.request().getHeader("login"))
+                    .put("canCreate", canCreate.succeeded() && canCreate.result())
+                    .put("canDelete", canDelete.succeeded() && canDelete.result())
+                    .put("canUpdate", canUpdate.succeeded() && canUpdate.result()),
+                  new JWTOptions()
+                    .setSubject("Wiki API")
+                    .setIssuer("Vert.x"));
+                context.response().putHeader("Content-Type", "text/plain").end(token);
+              });
+            });
+          });
+        } else {
+          context.fail(401);
+        }
+      });
+    });
+
     apiRouter.get("/pages").handler(this::apiRoot);
     apiRouter.get("/pages/:id").handler(this::apiGetPage);
     apiRouter.post().handler(BodyHandler.create());
@@ -135,10 +179,14 @@ public class HttpServerVerticle extends AbstractVerticle {
   }
 
   private void apiDeletePage(RoutingContext context) {
-    int id = Integer.valueOf(context.request().getParam("id"));
-    dbService.deletePage(id, reply -> {
-      handleSimpleDbReply(context, reply);
-    });
+    if (context.user().principal().getBoolean("canDelete", false)) {
+      int id = Integer.valueOf(context.request().getParam("id"));
+      dbService.deletePage(id, reply -> {
+        handleSimpleDbReply(context, reply);
+      });
+    } else {
+      context.fail(401);
+    }
   }
 
   private void handleSimpleDbReply(RoutingContext context, AsyncResult<Void> reply) {
@@ -156,14 +204,18 @@ public class HttpServerVerticle extends AbstractVerticle {
   }
 
   private void apiUpdatePage(RoutingContext context) {
-    int id = Integer.valueOf(context.request().getParam("id"));
-    JsonObject page = context.getBodyAsJson();
-    if (!validateJsonPageDocument(context, page, "markdown")) {
-      return;
+    if (context.user().principal().getBoolean("canUpdate", false)) {
+      int id = Integer.valueOf(context.request().getParam("id"));
+      JsonObject page = context.getBodyAsJson();
+      if (!validateJsonPageDocument(context, page, "markdown")) {
+        return;
+      }
+      dbService.savePage(id, page.getString("markdown"), reply -> {
+        handleSimpleDbReply(context, reply);
+      });
+    } else {
+      context.fail(401);
     }
-    dbService.savePage(id, page.getString("markdown"), reply -> {
-      handleSimpleDbReply(context, reply);
-    });
   }
 
   private boolean validateJsonPageDocument(RoutingContext context, JsonObject page, String... expectedKeys) {
@@ -180,23 +232,27 @@ public class HttpServerVerticle extends AbstractVerticle {
   }
 
   private void apiCreatePage(RoutingContext context) {
-    JsonObject page = context.getBodyAsJson();
-    if (!validateJsonPageDocument(context, page, "name", "markdown")) {
-      return;
-    }
-    dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
-      if (reply.succeeded()) {
-        context.response().setStatusCode(201);
-        context.response().putHeader("Content-Type", "application/json");
-        context.response().end(new JsonObject().put("success", true).encode());
-      } else {
-        context.response().setStatusCode(500);
-        context.response().putHeader("Content-Type", "application/json");
-        context.response().end(new JsonObject()
-          .put("success", false)
-          .put("error", reply.cause().getMessage()).encode());
+    if (context.user().principal().getBoolean("canCreate", false)) {
+      JsonObject page = context.getBodyAsJson();
+      if (!validateJsonPageDocument(context, page, "name", "markdown")) {
+        return;
       }
-    });
+      dbService.createPage(page.getString("name"), page.getString("markdown"), reply -> {
+        if (reply.succeeded()) {
+          context.response().setStatusCode(201);
+          context.response().putHeader("Content-Type", "application/json");
+          context.response().end(new JsonObject().put("success", true).encode());
+        } else {
+          context.response().setStatusCode(500);
+          context.response().putHeader("Content-Type", "application/json");
+          context.response().end(new JsonObject()
+            .put("success", false)
+            .put("error", reply.cause().getMessage()).encode());
+        }
+      });
+    } else {
+      context.fail(401);
+    }
   }
 
   private void apiGetPage(RoutingContext context) {
